% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{dropbox_clear_token}
\alias{dropbox_clear_token}
\title{Clear cached Dropbox token}
\usage{
dropbox_clear_token(cache_path = "~/R/.dropbox_token.rds")
}
\arguments{
\item{cache_path}{Path to cached token}
}
\description{
Clear cached Dropbox token
}
